requirements.txt
fastapi==0.120.3
psycopg2-binary==2.9.10
pydantic==2.12.3
pydantic[email]==2.12.3
python-dotenv==1.2.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4          
bcrypt==4.0.1                  
openai
requests
asyncpg
